+++ src/proto.rs
@@ -6,7 +6,7 @@ pub mod message;
 pub mod privatekey;
 pub mod signature;
 
-pub use self::error::*;
+pub use self::error::{ProtoError as Error, ProtoResult as Result, *};
 pub use self::message::*;
 pub use self::privatekey::*;
 pub use self::signature::*;
