+++ src/proto/message.rs
@@ -1,588 +1,20 @@
 //! Agent protocol message structures.
 
-pub mod add_identity;
-pub mod identity;
-pub mod sign;
-pub mod unparsed;
+mod add_remove;
+mod extension;
+mod identity;
+mod request;
+mod response;
+mod sign;
+mod unparsed;
 
-use ssh_encoding::{CheckedSum, Decode, Encode, Error as EncodingError, Reader, Writer};
-use ssh_key::{public::KeyData, Error, Signature};
-
-pub use self::{add_identity::*, identity::*, sign::*, unparsed::*};
-use super::{
-    extension::{KeyConstraintExtension, MessageExtension},
-    ProtoError,
+pub use self::{
+    add_remove::*, extension::*, identity::*, request::*, response::*, sign::*, unparsed::*,
 };
-
-type Result<T> = core::result::Result<T, ProtoError>;
-
-/// Remove a key from an agent.
-///
-/// This structure is sent in a [`Request::RemoveIdentity`] (`SSH_AGENTC_REMOVE_IDENTITY`) message.
-///
-/// Described in [draft-miller-ssh-agent-14 § 3.4](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3.4)
-#[derive(Clone, PartialEq, Debug)]
-pub struct RemoveIdentity {
-    /// The public key portion of the [`Identity`] to be removed
-    pub pubkey: KeyData,
-}
-
-impl Decode for RemoveIdentity {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let pubkey = reader.read_prefixed(KeyData::decode)?;
-
-        Ok(Self { pubkey })
-    }
-}
-
-impl Encode for RemoveIdentity {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        self.pubkey.encoded_len_prefixed()
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        self.pubkey.encode_prefixed(writer)
-    }
-}
-
-/// Pointer to a key in a hardware token, along with an optional PIN.
-///
-/// This structure is sent in a [`Request::AddSmartcardKey`] (`SSH_AGENTC_ADD_SMARTCARD_KEY`) message.
-///
-/// Described in [draft-miller-ssh-agent-14 § 3.2](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3.2)
-#[derive(Clone, PartialEq, Debug)]
-pub struct SmartcardKey {
-    /// An opaque identifier for the hardware token
-    ///
-    /// Note: the interpretation of "id" is not defined by the protocol,
-    /// but is left solely up to the agent.
-    pub id: String,
-
-    /// An optional password to unlock the key
-    pub pin: String,
-}
-
-impl Decode for SmartcardKey {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let id = String::decode(reader)?;
-        let pin = String::decode(reader)?;
-
-        Ok(Self { id, pin })
-    }
-}
-
-impl Encode for SmartcardKey {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        [self.id.encoded_len()?, self.pin.encoded_len()?].checked_sum()
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        self.id.encode(writer)?;
-        self.pin.encode(writer)?;
-
-        Ok(())
-    }
-}
-
-/// A key constraint, used to place limitations on how and where a key can be used.
-///
-/// Key constraints are set along with a key when are added to an agent.
-///
-/// Specifically, they appear in special `SSH_AGENTC_ADD_*` message variants:
-/// - [`Request::AddIdConstrained`]
-/// - [`Request::AddSmartcardKeyConstrained`]
-#[derive(Clone, PartialEq, Debug)]
-pub enum KeyConstraint {
-    /// Limit the key's lifetime by deleting it after the specified duration (in seconds)
-    Lifetime(u32),
-
-    /// Require explicit user confirmation for each private key operation using the key.
-    Confirm,
-
-    /// Experimental or private-use constraints
-    ///
-    /// Contains:
-    /// - An extension name indicating the type of the constraint (as a UTF-8 string).
-    /// - Extension-specific content
-    ///
-    /// Extension names should be suffixed by the implementation domain
-    /// as per [RFC4251 § 4.2](https://www.rfc-editor.org/rfc/rfc4251.html#section-4.2),
-    /// e.g. "foo@example.com"
-    Extension(Extension),
-}
-
-impl Decode for KeyConstraint {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let constraint_type = u8::decode(reader)?;
-        // see: https://www.ietf.org/archive/id/draft-miller-ssh-agent-12.html#section-5.2
-        Ok(match constraint_type {
-            1 => KeyConstraint::Lifetime(u32::decode(reader)?),
-            2 => KeyConstraint::Confirm,
-            255 => {
-                let name = String::decode(reader)?;
-                let details: Vec<u8> = Vec::decode(reader)?;
-                KeyConstraint::Extension(Extension {
-                    name,
-                    details: Unparsed::from(details),
-                })
-            }
-            _ => return Err(Error::AlgorithmUnknown)?, // FIXME: it should be our own type
-        })
-    }
-}
-
-impl Encode for KeyConstraint {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        let base = u8::MAX.encoded_len()?;
-
-        match self {
-            Self::Lifetime(lifetime) => base
-                .checked_add(lifetime.encoded_len()?)
-                .ok_or(EncodingError::Length),
-            Self::Confirm => Ok(base),
-            Self::Extension(extension) => [
-                base,
-                extension.name.encoded_len()?,
-                extension.details.encoded_len_prefixed()?,
-            ]
-            .checked_sum(),
-        }
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        match self {
-            Self::Lifetime(lifetime) => {
-                1u8.encode(writer)?;
-                lifetime.encode(writer)
-            }
-            Self::Confirm => 2u8.encode(writer),
-            Self::Extension(extension) => {
-                255u8.encode(writer)?;
-                extension.name.encode(writer)?;
-                extension.details.encode_prefixed(writer)
-            }
-        }
-    }
-}
-
-/// Add a key in a hardware token to an agent, with constraints on it's use.
-///
-/// This structure is sent in a [`Request::AddSmartcardKeyConstrained`] (`SSH_AGENTC_ADD_SMARTCARD_KEY_CONSTRAINED`) message.
-///
-/// This is a variant of [`Request::AddSmartcardKey`] with a set of [`KeyConstraint`]s attached.
-///
-/// Described in [draft-miller-ssh-agent-14 § 3.2.6](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3.2.6)
-#[derive(Clone, PartialEq, Debug)]
-pub struct AddSmartcardKeyConstrained {
-    /// A key stored on a hardware token.
-    pub key: SmartcardKey,
-
-    /// Constraints to be placed on the `key`.
-    pub constraints: Vec<KeyConstraint>,
-}
-
-impl Decode for AddSmartcardKeyConstrained {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let key = SmartcardKey::decode(reader)?;
-        let mut constraints = vec![];
-
-        while !reader.is_finished() {
-            constraints.push(KeyConstraint::decode(reader)?);
-        }
-        Ok(Self { key, constraints })
-    }
-}
-
-impl Encode for AddSmartcardKeyConstrained {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        self.constraints
-            .iter()
-            .try_fold(self.key.encoded_len()?, |acc, e| {
-                let constraint_len = e.encoded_len()?;
-                usize::checked_add(acc, constraint_len).ok_or(EncodingError::Length)
-            })
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        self.key.encode(writer)?;
-        for constraint in &self.constraints {
-            constraint.encode(writer)?;
-        }
-        Ok(())
-    }
-}
-
-/// Container for SSH agent protocol extension messages
-///
-/// This structure is sent as part of a [`Request::Extension`] (`SSH_AGENT_EXTENSION_RESPONSE`) message.
-///
-/// Described in [draft-miller-ssh-agent-14 § 3.8](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3.8).
-#[derive(Clone, PartialEq, Debug)]
-pub struct Extension {
-    /// Indicates the type of the extension message (as a UTF-8 string)
-    ///
-    /// Extension names should be suffixed by the implementation domain
-    /// as per [RFC4251 § 4.2](https://www.rfc-editor.org/rfc/rfc4251.html#section-4.2),
-    /// e.g. "foo@example.com"
-    pub name: String,
-
-    /// Extension-specific content
-    pub details: Unparsed,
-}
-
-impl Extension {
-    /// Create a new [`Extension`] from a [`MessageExtension`]
-    /// structure implementing [`ssh_encoding::Encode`]
-    pub fn new_message<T>(extension: T) -> Result<Self>
-    where
-        T: MessageExtension + Encode,
-    {
-        let mut buffer: Vec<u8> = vec![];
-        extension.encode(&mut buffer)?;
-        Ok(Self {
-            name: T::NAME.into(),
-            details: buffer.into(),
-        })
-    }
-
-    /// Attempt to parse a an extension object into a
-    /// [`MessageExtension`] structure
-    /// implementing [`ssh_encoding::Decode`].
-    ///
-    /// If there is a mismatch between the extension name
-    /// and the [`MessageExtension::NAME`], this method
-    /// will return [`None`]
-    pub fn parse_message<T>(&self) -> std::result::Result<Option<T>, <T as Decode>::Error>
-    where
-        T: MessageExtension + Decode,
-    {
-        if T::NAME == self.name {
-            Ok(Some(self.details.parse::<T>()?))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Create a new [`Extension`] from a [`KeyConstraintExtension`]
-    /// structure implementing [`ssh_encoding::Encode`]
-    pub fn new_key_constraint<T>(extension: T) -> Result<Self>
-    where
-        T: KeyConstraintExtension + Encode,
-    {
-        let mut buffer: Vec<u8> = vec![];
-        extension.encode(&mut buffer)?;
-        Ok(Self {
-            name: T::NAME.into(),
-            details: buffer.into(),
-        })
-    }
-
-    /// Attempt to parse a an extension object into a
-    /// [`KeyConstraintExtension`] structure
-    /// implementing [`ssh_encoding::Decode`].
-    ///
-    /// If there is a mismatch between the extension name
-    /// and the [`KeyConstraintExtension::NAME`], this method
-    /// will return [`None`]
-    pub fn parse_key_constraint<T>(&self) -> std::result::Result<Option<T>, <T as Decode>::Error>
-    where
-        T: KeyConstraintExtension + Decode,
-    {
-        if T::NAME == self.name {
-            Ok(Some(self.details.parse::<T>()?))
-        } else {
-            Ok(None)
-        }
-    }
-}
-
-impl Decode for Extension {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let name = String::decode(reader)?;
-        let mut details = vec![0; reader.remaining_len()];
-        reader.read(&mut details)?;
-        Ok(Self {
-            name,
-            details: details.into(),
-        })
-    }
-}
-
-impl Encode for Extension {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        [self.name.encoded_len()?, self.details.encoded_len()?].checked_sum()
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        self.name.encode(writer)?;
-        self.details.encode(writer)?;
-        Ok(())
-    }
-}
-
-/// SSH agent protocol request messages.
-///
-/// These message types are sent from a client *to* an agent.
-///
-/// Described in [draft-miller-ssh-agent-14 § 3](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3).
-#[derive(Clone, PartialEq, Debug)]
-pub enum Request {
-    /// Request a list of all identities (public key/certificate & comment)
-    /// from an agent
-    RequestIdentities,
-
-    /// Perform a private key signature operation using a key
-    /// stored in the agent
-    SignRequest(SignRequest),
-
-    /// Add an identity (private key/certificate & comment) to an agent
-    AddIdentity(AddIdentity),
-
-    /// Remove an identity from an agent
-    RemoveIdentity(RemoveIdentity),
-
-    /// Remove all identities from an agent
-    RemoveAllIdentities,
-
-    /// Add an identity (private key/certificate & comment) to an agent
-    /// where the private key is stored on a hardware token
-    AddSmartcardKey(SmartcardKey),
-
-    /// Remove a key stored on a hardware token from an agent
-    RemoveSmartcardKey(SmartcardKey),
-
-    /// Temporarily lock an agent with a pass-phrase
-    Lock(String),
-
-    /// Unlock a locked agaent with a pass-phrase
-    Unlock(String),
-
-    /// Add an identity (private key/certificate & comment) to an agent,
-    /// with constraints on it's usage
-    AddIdConstrained(AddIdentityConstrained),
-
-    /// Add an identity (private key/certificate & comment) to an agent
-    /// where the private key is stored on a hardware token,
-    /// with constraints on it's usage
-    AddSmartcardKeyConstrained(AddSmartcardKeyConstrained),
-
-    /// Send a vendor-specific message via the agent protocol,
-    /// identified by an *extension type*.
-    Extension(Extension),
-}
-
-/// SSH agent protocol response messages.
-///
-/// These message types are sent to a client *from* an agent (in response to a [`Request`] message).
-///
-/// Described in [draft-miller-ssh-agent-14 § 3](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-3).
-#[derive(Clone, PartialEq, Debug)]
-pub enum Response {
-    /// Indicates generic agent failure
-    Failure,
-
-    /// Indicates generic agent success
-    Success,
-
-    /// A list of identities, sent in response to
-    /// a [`Request::RequestIdentities`] message.
-    IdentitiesAnswer(Vec<Identity>),
-
-    /// A signature, sent in response to
-    /// a [`Request::SignRequest`] message.
-    SignResponse(Signature),
-
-    /// Indicates generic extension failure
-    ExtensionFailure,
-
-    /// Send a vendor-specific response message via the agent protocol,
-    /// identified by an *extension type*.
-    ExtensionResponse(Extension),
-}
-
-impl Request {
-    /// The protocol message identifier for a given [`Request`] message type.
-    ///
-    /// Described in [draft-miller-ssh-agent-14 § 6.1](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-6.1).
-    pub fn message_id(&self) -> u8 {
-        match self {
-            Self::RequestIdentities => 11,
-            Self::SignRequest(_) => 13,
-            Self::AddIdentity(_) => 17,
-            Self::RemoveIdentity(_) => 18,
-            Self::RemoveAllIdentities => 19,
-            Self::AddSmartcardKey(_) => 20,
-            Self::RemoveSmartcardKey(_) => 21,
-            Self::Lock(_) => 22,
-            Self::Unlock(_) => 23,
-            Self::AddIdConstrained(_) => 25,
-            Self::AddSmartcardKeyConstrained(_) => 26,
-            Self::Extension(_) => 27,
-        }
-    }
-}
-
-impl Response {
-    /// The protocol message identifier for a given [`Response`] message type.
-    ///
-    /// Described in [draft-miller-ssh-agent-14 § 6.1](https://www.ietf.org/archive/id/draft-miller-ssh-agent-14.html#section-6.1).
-    pub fn message_id(&self) -> u8 {
-        match self {
-            Self::Failure => 5,
-            Self::Success => 6,
-            Self::IdentitiesAnswer(_) => 12,
-            Self::SignResponse(_) => 14,
-            Self::ExtensionFailure => 28,
-            Self::ExtensionResponse(_) => 29,
-        }
-    }
-}
-
-impl Decode for Request {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let message_type = u8::decode(reader)?;
-
-        match message_type {
-            11 => Ok(Self::RequestIdentities),
-            13 => SignRequest::decode(reader).map(Self::SignRequest),
-            17 => AddIdentity::decode(reader).map(Self::AddIdentity),
-            18 => RemoveIdentity::decode(reader).map(Self::RemoveIdentity),
-            19 => Ok(Self::RemoveAllIdentities),
-            20 => SmartcardKey::decode(reader).map(Self::AddSmartcardKey),
-            21 => SmartcardKey::decode(reader).map(Self::RemoveSmartcardKey),
-            22 => Ok(String::decode(reader).map(Self::Lock)?),
-            23 => Ok(String::decode(reader).map(Self::Unlock)?),
-            25 => AddIdentityConstrained::decode(reader).map(Self::AddIdConstrained),
-            26 => AddSmartcardKeyConstrained::decode(reader).map(Self::AddSmartcardKeyConstrained),
-            27 => Extension::decode(reader).map(Self::Extension),
-            command => Err(ProtoError::UnsupportedCommand { command }),
-        }
-    }
-}
-
-impl Decode for Response {
-    type Error = ProtoError;
-
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let message_type = u8::decode(reader)?;
-
-        match message_type {
-            5 => Ok(Self::Failure),
-            6 => Ok(Self::Success),
-            12 => Identity::decode_vec(reader).map(Self::IdentitiesAnswer),
-            14 => {
-                Ok(reader
-                    .read_prefixed(|reader| Signature::decode(reader).map(Self::SignResponse))?)
-            }
-            28 => Ok(Self::ExtensionFailure),
-            29 => Extension::decode(reader).map(Self::ExtensionResponse),
-            command => Err(ProtoError::UnsupportedCommand { command }),
-        }
-    }
-}
-
-impl Encode for Request {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        let message_id_len = 1;
-        let payload_len = match self {
-            Self::RequestIdentities => 0,
-            Self::SignRequest(request) => request.encoded_len()?,
-            Self::AddIdentity(identity) => identity.encoded_len()?,
-            Self::RemoveIdentity(identity) => identity.encoded_len()?,
-            Self::RemoveAllIdentities => 0,
-            Self::AddSmartcardKey(key) => key.encoded_len()?,
-            Self::RemoveSmartcardKey(key) => key.encoded_len()?,
-            Self::Lock(passphrase) => passphrase.encoded_len()?,
-            Self::Unlock(passphrase) => passphrase.encoded_len()?,
-            Self::AddIdConstrained(key) => key.encoded_len()?,
-            Self::AddSmartcardKeyConstrained(key) => key.encoded_len()?,
-            Self::Extension(extension) => extension.encoded_len()?,
-        };
-
-        [message_id_len, payload_len].checked_sum()
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        let message_id: u8 = self.message_id();
-        message_id.encode(writer)?;
-
-        match self {
-            Self::RequestIdentities => {}
-            Self::SignRequest(request) => request.encode(writer)?,
-            Self::AddIdentity(identity) => identity.encode(writer)?,
-            Self::RemoveIdentity(identity) => identity.encode(writer)?,
-            Self::RemoveAllIdentities => {}
-            Self::AddSmartcardKey(key) => key.encode(writer)?,
-            Self::RemoveSmartcardKey(key) => key.encode(writer)?,
-            Self::Lock(passphrase) => passphrase.encode(writer)?,
-            Self::Unlock(passphrase) => passphrase.encode(writer)?,
-            Self::AddIdConstrained(identity) => identity.encode(writer)?,
-            Self::AddSmartcardKeyConstrained(key) => key.encode(writer)?,
-            Self::Extension(extension) => extension.encode(writer)?,
-        };
-
-        Ok(())
-    }
-}
-
-impl Encode for Response {
-    fn encoded_len(&self) -> ssh_encoding::Result<usize> {
-        let message_id_len = 1;
-        let payload_len = match self {
-            Self::Failure => 0,
-            Self::Success => 0,
-            Self::IdentitiesAnswer(ids) => {
-                let mut lengths = Vec::with_capacity(1 + ids.len());
-                // Prefixed length
-                lengths.push(4);
-
-                for id in ids {
-                    lengths.push(id.encoded_len()?);
-                }
-
-                lengths.checked_sum()?
-            }
-            Self::SignResponse(response) => response.encoded_len_prefixed()?,
-            Self::ExtensionFailure => 0,
-            Self::ExtensionResponse(extension) => extension.encoded_len()?,
-        };
-
-        [message_id_len, payload_len].checked_sum()
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> ssh_encoding::Result<()> {
-        let message_id: u8 = self.message_id();
-        message_id.encode(writer)?;
-
-        match self {
-            Self::Failure => {}
-            Self::Success => {}
-            Self::IdentitiesAnswer(ids) => {
-                (ids.len() as u32).encode(writer)?;
-                for id in ids {
-                    id.encode(writer)?;
-                }
-            }
-            Self::SignResponse(response) => response.encode_prefixed(writer)?,
-            Self::ExtensionFailure => {}
-            Self::ExtensionResponse(extension) => extension.encode(writer)?,
-        };
-
-        Ok(())
-    }
-}
+#[doc(hidden)]
+/// For compatibility with pre-0.5.0 type alias in this module
+/// that duplicated crate::proto::error::ProtoResult
+pub use super::Result;
 
 #[cfg(test)]
 mod tests {
@@ -594,15 +26,18 @@ mod tests {
         elliptic_curve::{bigint::Uint, ScalarPrimitive},
         EncodedPoint,
     };
+    use ssh_encoding::{Decode, Encode};
     use ssh_key::{
-        private::{EcdsaKeypair, EcdsaPrivateKey, KeypairData, RsaPrivateKey},
+        private::{EcdsaKeypair, EcdsaPrivateKey, RsaPrivateKey},
+        public::KeyData,
         Algorithm, Certificate, Mpint,
     };
 
     use super::*;
     use crate::proto::PrivateKeyData;
 
-    fn demo_key() -> EcdsaKeypair {
+    // Expose this function to all tests under `crate::proto::message`
+    pub(super) fn demo_key() -> EcdsaKeypair {
         EcdsaKeypair::NistP256 {
             public: EncodedPoint::from_affine_coordinates(
                 &hex!("cb244fcdb89de95bc8fd766e6b139abfc2649fb063b6c5e5a939e067e2a0d215").into(),
@@ -616,203 +51,6 @@ mod tests {
                 .unwrap(),
             )),
         }
-    }
-
-    #[test]
-    fn test_add_identity_constrained() {
-        let msg: &[u8] = &hex!(
-            "
-                        00 0000 1365 6364 7361 2d73
-            6861 322d 6e69 7374 7032 3536 0000 0008
-            6e69 7374 7032 3536 0000 0041 04cb 244f
-            cdb8 9de9 5bc8 fd76 6e6b 139a bfc2 649f
-            b063 b6c5 e5a9 39e0 67e2 a0d2 150a 660d
-            aca7 8f6c 24a0 4253 73d6 ea83 e36f 8a1f
-            8b82 8a60 e77a 97a9 441b cc09 8700 0000
-            2100 ffd9 f2ce 4d0e e587 0d8d c7cf 771a
-            7669 a0b9 6fe4 4bb5 8a8a 0bc7 5a76 b4f7
-            8240 0000 000c 6261 6c6f 6f40 616e 6765
-            6c61 0100 0000 02
-        "
-        );
-        let mut reader = msg;
-
-        let out = AddIdentityConstrained::decode(&mut reader).unwrap();
-
-        let expected = AddIdentityConstrained {
-            identity: AddIdentity {
-                credential: Credential::Key {
-                    privkey: KeypairData::Ecdsa(demo_key()),
-                    comment: "baloo@angela".to_string(),
-                },
-            },
-            constraints: vec![KeyConstraint::Lifetime(2)],
-        };
-        assert_eq!(out, expected);
-
-        let mut buf = vec![];
-        expected.encode(&mut buf).expect("serialize message");
-        assert_eq!(expected.encoded_len().expect("len message"), buf.len());
-        assert_eq!(buf, msg);
-
-        let msg: &[u8] = &hex!(
-            "
-                        00 0000 1365 6364 7361 2d73
-            6861 322d 6e69 7374 7032 3536 0000 0008
-            6e69 7374 7032 3536 0000 0041 04cb 244f
-            cdb8 9de9 5bc8 fd76 6e6b 139a bfc2 649f
-            b063 b6c5 e5a9 39e0 67e2 a0d2 150a 660d
-            aca7 8f6c 24a0 4253 73d6 ea83 e36f 8a1f
-            8b82 8a60 e77a 97a9 441b cc09 8700 0000
-            2100 ffd9 f2ce 4d0e e587 0d8d c7cf 771a
-            7669 a0b9 6fe4 4bb5 8a8a 0bc7 5a76 b4f7
-            8240 0000 000c 6261 6c6f 6f40 616e 6765
-            6c61 ff00 0000 2472 6573 7472 6963 742d
-            6465 7374 696e 6174 696f 6e2d 7630 3040
-            6f70 656e 7373 682e 636f 6d00 0002 7300
-            0002 6f00 0000 0c00 0000 0000 0000 0000
-            0000 0000 0002 5700 0000 0000 0000 0a67
-            6974 6875 622e 636f 6d00 0000 0000 0000
-            3300 0000 0b73 7368 2d65 6432 3535 3139
-            0000 0020 e32a aa79 15ce b9b4 49d1 ba50
-            ea2a 28bb 1a6e 01f9 0bda 245a 2d1d 8769
-            7d18 a265 0000 0001 9700 0000 0773 7368
-            2d72 7361 0000 0003 0100 0100 0001 8100
-            a3ee 774d c50a 3081 c427 8ec8 5c2e ba8f
-            1228 a986 7b7e 5534 ef0c fea6 1c12 fd8f
-            568d 5246 3851 ed60 bf09 c62d 594e 8467
-            98ae 765a 3204 4aeb e3ca 0945 da0d b0bb
-            aad6 d6f2 0224 84be da18 2b0e aff0 b9e9
-            224c cbf0 4265 fc5d d675 b300 ec52 0cf8
-            15b2 67ab 3816 1f36 a96d 57df e158 2a81
-            cb02 0d21 1fb9 7488 3a25 327b da97 04a4
-            48dc 6205 e413 6604 1575 7524 79ec 2a06
-            cb58 d961 49ca 9bd9 49b2 4644 32ca d44b
-            b4bf b7f1 31b1 9310 9f96 63be e59f 0249
-            2358 ec68 9d8c c219 ed0e 3332 3036 9f59
-            c6ae 54c3 933c 030a cc3e c2a1 4f19 0035
-            efd7 277c 658e 5915 6bba 3d7a cfa5 f2bf
-            1be3 2706 f3d3 0419 ef95 cae6 d292 6fb1
-            4dc9 e204 b384 d3e2 393e 4b87 613d e014
-            0b9c be6c 3622 ad88 0ce0 60bb b849 f3b6
-            7672 6955 90ec 1dfc d402 b841 daf0 b79d
-            59a8 4c4a 6d0a 5350 d9fe 123a a84f 0bea
-            363e 24ab 1e50 5022 344e 14bf 6243 b124
-            25e6 3d45 996e 18e9 0a0e 7a8b ed9a 07a0
-            a62b 6246 867e 7b2b 99a3 d0c3 5d05 7038
-            fd69 f01f a5e8 3d62 732b 9372 bb6c c1de
-            7019 a7e4 b986 942c fa9d 6f37 5ff0 b239
-            0000 0000 6800 0000 1365 6364 7361 2d73
-            6861 322d 6e69 7374 7032 3536 0000 0008
-            6e69 7374 7032 3536 0000 0041 0449 8a48
-            4363 4047 b33a 6c64 64cc bba2 92a0 c050
-            7d9e 4b79 611a d832 336e 1b93 7cee e460
-            83a0 8bad ba39 c007 53ff 2eaf d262 95d1
-            4db0 d166 7660 1ffe f93a 6872 4800 0000
-            0000"
-        );
-        let mut reader = msg;
-
-        let out = AddIdentityConstrained::decode(&mut reader).unwrap();
-
-        let expected = AddIdentityConstrained {
-            identity: AddIdentity {
-                credential: Credential::Key {
-                    privkey: KeypairData::Ecdsa(demo_key()),
-                    comment: "baloo@angela".to_string(),
-                },
-            },
-            constraints: vec![KeyConstraint::Extension(Extension {
-                name: "restrict-destination-v00@openssh.com".to_string(),
-                details: Unparsed::from(
-                    hex!(
-                        "                                    00
-                        0002 6f00 0000 0c00 0000 0000 0000 0000
-                        0000 0000 0002 5700 0000 0000 0000 0a67
-                        6974 6875 622e 636f 6d00 0000 0000 0000
-                        3300 0000 0b73 7368 2d65 6432 3535 3139
-                        0000 0020 e32a aa79 15ce b9b4 49d1 ba50
-                        ea2a 28bb 1a6e 01f9 0bda 245a 2d1d 8769
-                        7d18 a265 0000 0001 9700 0000 0773 7368
-                        2d72 7361 0000 0003 0100 0100 0001 8100
-                        a3ee 774d c50a 3081 c427 8ec8 5c2e ba8f
-                        1228 a986 7b7e 5534 ef0c fea6 1c12 fd8f
-                        568d 5246 3851 ed60 bf09 c62d 594e 8467
-                        98ae 765a 3204 4aeb e3ca 0945 da0d b0bb
-                        aad6 d6f2 0224 84be da18 2b0e aff0 b9e9
-                        224c cbf0 4265 fc5d d675 b300 ec52 0cf8
-                        15b2 67ab 3816 1f36 a96d 57df e158 2a81
-                        cb02 0d21 1fb9 7488 3a25 327b da97 04a4
-                        48dc 6205 e413 6604 1575 7524 79ec 2a06
-                        cb58 d961 49ca 9bd9 49b2 4644 32ca d44b
-                        b4bf b7f1 31b1 9310 9f96 63be e59f 0249
-                        2358 ec68 9d8c c219 ed0e 3332 3036 9f59
-                        c6ae 54c3 933c 030a cc3e c2a1 4f19 0035
-                        efd7 277c 658e 5915 6bba 3d7a cfa5 f2bf
-                        1be3 2706 f3d3 0419 ef95 cae6 d292 6fb1
-                        4dc9 e204 b384 d3e2 393e 4b87 613d e014
-                        0b9c be6c 3622 ad88 0ce0 60bb b849 f3b6
-                        7672 6955 90ec 1dfc d402 b841 daf0 b79d
-                        59a8 4c4a 6d0a 5350 d9fe 123a a84f 0bea
-                        363e 24ab 1e50 5022 344e 14bf 6243 b124
-                        25e6 3d45 996e 18e9 0a0e 7a8b ed9a 07a0
-                        a62b 6246 867e 7b2b 99a3 d0c3 5d05 7038
-                        fd69 f01f a5e8 3d62 732b 9372 bb6c c1de
-                        7019 a7e4 b986 942c fa9d 6f37 5ff0 b239
-                        0000 0000 6800 0000 1365 6364 7361 2d73
-                        6861 322d 6e69 7374 7032 3536 0000 0008
-                        6e69 7374 7032 3536 0000 0041 0449 8a48
-                        4363 4047 b33a 6c64 64cc bba2 92a0 c050
-                        7d9e 4b79 611a d832 336e 1b93 7cee e460
-                        83a0 8bad ba39 c007 53ff 2eaf d262 95d1
-                        4db0 d166 7660 1ffe f93a 6872 4800 0000
-                        0000"
-                    )
-                    .to_vec(),
-                ),
-            })],
-        };
-
-        assert_eq!(out, expected);
-
-        let mut buf = vec![];
-        expected.encode(&mut buf).expect("serialize message");
-        assert_eq!(expected.encoded_len().expect("len message"), buf.len());
-        assert_eq!(buf, msg);
-    }
-
-    #[test]
-    fn test_add_identity() {
-        let msg: &[u8] = &hex!(
-            "
-	                00 0000 1365 6364 7361 2d73
-	    6861 322d 6e69 7374 7032 3536 0000 0008
-	    6e69 7374 7032 3536 0000 0041 04cb 244f
-	    cdb8 9de9 5bc8 fd76 6e6b 139a bfc2 649f
-	    b063 b6c5 e5a9 39e0 67e2 a0d2 150a 660d
-	    aca7 8f6c 24a0 4253 73d6 ea83 e36f 8a1f
-	    8b82 8a60 e77a 97a9 441b cc09 8700 0000
-	    2100 ffd9 f2ce 4d0e e587 0d8d c7cf 771a
-	    7669 a0b9 6fe4 4bb5 8a8a 0bc7 5a76 b4f7
-	    8240 0000 000c 6261 6c6f 6f40 616e 6765
-	    6c61
-        "
-        );
-        let mut reader = msg;
-
-        let out = AddIdentity::decode(&mut reader).expect("parse message");
-
-        let expected = AddIdentity {
-            credential: Credential::Key {
-                privkey: KeypairData::Ecdsa(demo_key()),
-                comment: "baloo@angela".to_string(),
-            },
-        };
-        assert_eq!(out, expected);
-
-        let mut buf = vec![];
-        expected.encode(&mut buf).expect("serialize message");
-        assert_eq!(buf, msg);
     }
 
     #[test]
